One step closer

+++ Assets/Scripts/Units/UnitStats.cs
@@ -244,7 +244,7 @@ public class UnitStats : MonoBehaviour
     public float GetMagicResist() => magicResistModifier.Calculate(data.magicResist);
     public float GetMoveSpeed() => speedModifier.Calculate(data.moveSpeed);
     public float GetAttackSpeed() => attackSpeedModifier.Calculate(data.attackSpeed);
-    public float GetRange() => rangeModifier.Calculate(data.range);
+    public int GetRange() => (int) rangeModifier.Calculate(data.range);
     public int GetDetectRange() => data.detectRange;
     public float GetLifesteal() => lifeStealModifier.Calculate(data.lifestealPercent);
     public float GetCriticalChance() => data.criticalChance;
